Internship Details
We need to show the internship what we have in the 7th Sem. I have attached the report & PPT of my internship which i have done. You need to follow the same format for making the PPT & Report. 
To make the flow chart which i have done in the (E-Commerce.png) you can follow the below link
https://whimsical.com